nucleus/requirements.txt
connexion==2.7.0
elasticsearch==7.10.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
prometheus-flask-exporter==0.18.1
psycopg2-binary==2.8.6
python-dotenv==0.15.0
python-jose==3.2.0
swagger-ui-bundle==0.0.8
Pillow==8.0.1
Faker==5.0.2
